implement service: friend, group, membership

+++ ngobrol-server/src/com/ngobrol/server/user/service/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
 
   }
 
-  public void sendToGroup(User sender, int groupId, String message) {
+  public void sendToGroup(User sender, long groupId, String message) {
 
   }
